Move shared code for services to ProxyExecutionService
