+++ packages/snaps-controllers/src/services/offscreen/OffscreenExecutionService.ts
@@ -1,21 +1,15 @@
 import { BrowserRuntimePostMessageStream } from '@metamask/post-message-stream';
-import { nanoid } from 'nanoid';
 
-import type { ExecutionServiceArgs, Job } from '../AbstractExecutionService';
-import { AbstractExecutionService } from '../AbstractExecutionService';
-import { ProxyPostMessageStream } from '../ProxyPostMessageStream';
+import type { ExecutionServiceArgs } from '../AbstractExecutionService';
+import { ProxyExecutionService } from '../proxy/ProxyExecutionService';
 
 type OffscreenExecutionEnvironmentServiceArgs = {
   documentUrl: URL;
   frameUrl: URL;
 } & ExecutionServiceArgs;
 
-export class OffscreenExecutionService extends AbstractExecutionService<string> {
+export class OffscreenExecutionService extends ProxyExecutionService {
   public readonly documentUrl: URL;
-
-  public readonly frameUrl: URL;
-
-  readonly #runtimeStream: BrowserRuntimePostMessageStream;
 
   /**
    * Create a new offscreen execution service.
@@ -40,32 +34,14 @@ export class OffscreenExecutionService extends AbstractExecutionService<string> 
     super({
       messenger,
       setupSnapProvider,
+      frameUrl,
+      stream: new BrowserRuntimePostMessageStream({
+        name: 'parent',
+        target: 'child',
+      }),
     });
 
     this.documentUrl = documentUrl;
-    this.frameUrl = frameUrl;
-    this.#runtimeStream = new BrowserRuntimePostMessageStream({
-      name: 'parent',
-      target: 'child',
-    });
-  }
-
-  /**
-   * Send a termination command to the offscreen document.
-   *
-   * @param job - The job to terminate.
-   */
-  protected async terminateJob(job: Job<string>) {
-    // The `AbstractExecutionService` will have already closed the job stream,
-    // so we write to the runtime stream directly.
-    this.#runtimeStream.write({
-      jobId: job.id,
-      data: {
-        jsonrpc: '2.0',
-        method: 'terminateJob',
-        id: nanoid(),
-      },
-    });
   }
 
   /**
@@ -76,20 +52,9 @@ export class OffscreenExecutionService extends AbstractExecutionService<string> 
    */
   protected async initEnvStream(jobId: string) {
     // Lazily create the offscreen document.
-    await this.createDocument();
+    await this.#createDocument();
 
-    const stream = new ProxyPostMessageStream({
-      stream: this.#runtimeStream,
-      extra: {
-        // TODO: Rather than injecting the frame URL here, we should come up
-        // with a better way to do this. The frame URL is needed to avoid hard
-        // coding it in the offscreen execution environment.
-        frameUrl: this.frameUrl.toString(),
-      },
-      jobId,
-    });
-
-    return { worker: jobId, stream };
+    return super.initEnvStream(jobId);
   }
 
   /**
@@ -97,7 +62,7 @@ export class OffscreenExecutionService extends AbstractExecutionService<string> 
    *
    * If the document already exists, this does nothing.
    */
-  private async createDocument() {
+  async #createDocument() {
     // Extensions can only have a single offscreen document.
     if (await chrome.offscreen.hasDocument()) {
       return;
